Fix a zipmap parsing bug
The zipmap entries allow to dynamically update the values, but not the
keys. The 'Free' entry may only be present before the value, not the
key.

This patch fixes the zipmap entry parsing so that it is aware of what is
a key and what is a value, rather than assuming they are uniform.