fix ziplist endianness bug
Ziplists with lengths greater than 63 bytes and up to 16383 bytes appear
not to be encoded as small endian like the rest. The original source at
https://github.com/sripathikrishnan/redis-rdb-tools/blob/master/docs/RDB_File_Format.textile
doesn't mention the endianness, so 'native' is being used. This will
default to 'big' on most architectures, maybe implode on clusters with
varying endianness.

The patch also includes a fix for an explicitly big endian format for
String values with length greater than or equal to 16384 bytes, as
showin at
https://github.com/sripathikrishnan/redis-rdb-tools/commit/4a4b7e073aae64a5ec96dc461a25983843f09ea1
in the original project.