Adding long heartbeat timeouts on connections
Redis will send a ping message roughly every 10 seconds, although this
feature appears undocumented. There have been cases where the connection
seems entirely blocked for no reason, which should prompt a reconnection
and a new sync.

This patch adds a heartbeat with a minute-long delay that will close the
connection and restart it, and ask to sync again if the server doesn't
answer to a PING command (with "+PONG\r\n") in the next minute.